+++ Package.swift
@@ -12,8 +12,13 @@ let package = Package(
         .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.4.3"),
     ],
     targets: [
+        .systemLibrary(
+            name: "CZlib",
+            path: "Sources/CZlib"
+        ),
         .target(
             name: "SwiftZIP",
+            dependencies: ["CZlib"],
             path: "Sources/SwiftZIP"
         ),
         .testTarget(